+++ packages/app/src/queries/songs.ts
@@ -2,9 +2,9 @@ import firebase from "firebase/app";
 import { clientStorage } from "../shared/utils";
 import { createQueryCache } from "../queries/cache";
 import { Song } from "../shared/types";
-import { storage, analytics, performance } from "../firebase";
+import { storage, withPerformanceAndAnalytics } from "../firebase";
 import { getDownloadURL } from "../storage";
-import { captureAndLogError, captureAndLog, withPerformanceAndAnalytics } from "../utils";
+import { captureAndLogError, captureAndLog } from "../utils";
 import { useUserData } from "../firestore";
 import { useMutation } from "react-query";
 import { useMemo } from "react";
