+++ packages/app/src/utils.ts
@@ -10,7 +10,6 @@ import {
 } from "react";
 import * as Sentry from "@sentry/browser";
 import tiny from "tinycolor2";
-import { performance, analytics } from "./firebase";
 
 export interface Disposer {
   dispose: () => void;
@@ -473,23 +472,4 @@ export const removeElementFromShuffled = <T>(
     mappingFrom: reverseMapping(newMappingTo),
     mappingTo: newMappingTo,
   };
-};
-
-export const withPerformanceAndAnalytics = <T>(
-  cb: () => Promise<T[]>,
-  name: string,
-) => async (): Promise<T[]> => {
-  const trace = performance.trace(name);
-  trace.start();
-
-  const result = await cb();
-
-  // If result errors out, the trace is never ended and nothing actually happens
-  trace.stop();
-  trace.putMetric("count", result.length);
-  analytics.logEvent(name, {
-    value: result.length,
-  });
-
-  return result;
 };
